+++ releasenotes.md
@@ -3,7 +3,8 @@
 ## 0.5.0 Change List
 
 ### Changes / improvements
-- '$$select' builtin for vector masked select.
+- `$$masked_load` / `$$masked_store` for vector masked load/store. 
+- `$$select` builtin for vector masked select.
 - Subtype matching in type switches.
 - Added parentof typeid property.
 - Slice assignment is expanded.